% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert2id}
\alias{convert2id}
\title{Convert a list of input taxa into taxon IDs}
\usage{
convert2id(taxonList)
}
\arguments{
\item{taxonList}{list of taxa}
}
\value{
A list of taxon IDs
}
\description{
Convert a list of input taxa into taxon IDs
}
\examples{
taxonList <- c("Homo sapiens", "5207", "40674", "4751")
taxFun:::convert2id(taxonList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
