% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert2name}
\alias{convert2name}
\title{Convert a list of input taxa into taxon names}
\usage{
convert2name(taxonList)
}
\arguments{
\item{taxonList}{list of taxa}
}
\value{
A list of taxon names
}
\description{
Convert a list of input taxa into taxon names
}
\examples{
taxonList <- c("Homo sapiens", "5207", "40674", "4751")
taxFun:::convert2name(taxonList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
