% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTaxonomyMatrix.R
\name{createTaxonomyMatrix}
\alias{createTaxonomyMatrix}
\title{Create an aligned taxonomy matrix for a list of taxon IDs}
\usage{
createTaxonomyMatrix(idList)
}
\arguments{
\item{idList}{list of taxonomy IDs}
}
\value{
A dataframe contains clustered taxonomy matrix
}
\description{
Create an aligned taxonomy matrix for a list of taxon IDs
}
\examples{
idList <- c("9606", "5207", "40674", "4751", "3702", "72664")
createTaxonomyMatrix(idList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
