% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{firstup}
\alias{firstup}
\title{Capitalize the first character of string(s)}
\usage{
firstup(x)
}
\arguments{
\item{x}{list of strings}
}
\value{
List of strings with the first character capitalized
}
\description{
Capitalize the first character of string(s)
}
\examples{
ll <- c("one", "tWO", "Three")
taxFun:::firstup(ll)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
