% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getAllTaxonomyRanks}
\alias{getAllTaxonomyRanks}
\title{Get all NCBI taxonomy rank names}
\usage{
getAllTaxonomyRanks()
}
\value{
A list of all available NCBI taxonomy rank names.
}
\description{
Get all NCBI taxonomy rank names
}
\examples{
getAllTaxonomyRanks()
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
