% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSampling.R
\name{getClade}
\alias{getClade}
\title{Get subset of taxa that belong to a specified taxonomy clade}
\usage{
getClade(taxonList = NULL, supertaxon = NULL)
}
\arguments{
\item{taxonList}{List of taxa}

\item{supertaxon}{Clade / supertaxon ID or name}
}
\value{
A dataframe contains sub-selected taxa including their IDs and names
}
\description{
Get subset of taxa that belong to a specified taxonomy clade
}
\examples{
taxonList <- c(
    214684, 246409, 3055, 336722, 3702, 400682, 441375, 45351, 4558, 559292, 
    6239, 7227, "Homo sapiens"
)
getClade(taxonList, supertaxon = "metazoa")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
