% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLCA.R
\name{getLCA}
\alias{getLCA}
\title{Get last common ancestor for a list of taxa}
\usage{
getLCA(taxonList = NULL)
}
\arguments{
\item{taxonList}{List of taxa}
}
\value{
A dataframe contains taxon ID and name of the last common ancestor
}
\description{
Get last common ancestor for a list of taxa
}
\examples{
taxonList <- c("Homo sapiens", "Mus musculus", "Bos taurus")
getLCA(taxonList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
