% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getPreTaxonomyFile}
\alias{getPreTaxonomyFile}
\title{Get PhyloProfile taxonomy file if available or create one if missing}
\usage{
getPreTaxonomyFile()
}
\value{
A data table contains the NCBI taxonomy DB from PhyloProfile
}
\description{
Get PhyloProfile taxonomy file if available or create one if missing
}
\examples{
preProcessedTaxonomy <- getPreTaxonomyFile()
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
