% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRank}
\alias{getRank}
\title{Get taxonomy ranks for a list of taxa}
\usage{
getRank(taxonList = NULL)
}
\arguments{
\item{taxonList}{list of taxa}
}
\value{
A dataframe contains input taxon Ids, their full names and the
corresponding ranks
}
\description{
Get taxonomy ranks for a list of taxa
}
\examples{
taxonList <- c("Homo sapiens", "5207", "40674", "4751")
getRank(taxonList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
