% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRanks}
\alias{getRanks}
\title{Get defined taxonomy ranks for a list of taxa}
\usage{
getRanks(taxonList = NULL, ranks = NULL)
}
\arguments{
\item{taxonList}{list of taxa}

\item{ranks}{list of defined ranks. If not given, all available ranks will
be considered}
}
\value{
A dataframe contains input taxon Ids, their full names and names of
defined taxonomy ranks
}
\description{
Get defined taxonomy ranks for a list of taxa
}
\examples{
taxonList <- c("Homo sapiens", "5207", "40674", "4751")
ranks <- c("family", "class", "phylum", "kingdom")
getRanks(taxonList, ranks)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
