% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRanks4Id}
\alias{getRanks4Id}
\title{Get defined taxonomy ranks for a list of taxon IDs}
\usage{
getRanks4Id(idList = NULL, ranks = NULL)
}
\arguments{
\item{idList}{list of taxonomy IDs}

\item{ranks}{list of defined ranks. If not given, all available ranks will
be considered}
}
\value{
A dataframe contains input taxon Ids, their full names and names of
defined taxonomy ranks
}
\description{
Get defined taxonomy ranks for a list of taxon IDs
}
\examples{
idList <- c("9606", "5207", "40674", "4751")
ranks <- c("family", "class", "phylum", "kingdom")
getRanks4Id(idList, ranks)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
