% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRanks4Name}
\alias{getRanks4Name}
\title{Get defined taxonomy ranks for a list of taxon names}
\usage{
getRanks4Name(nameList = NULL, ranks = NULL)
}
\arguments{
\item{nameList}{list of taxonomy names}

\item{ranks}{list of defined ranks. If not given, all available ranks will
be considered}
}
\value{
A dataframe contains input taxon names, their IDs and names of
defined taxonomy ranks
}
\description{
Get defined taxonomy ranks for a list of taxon names
}
\examples{
nameList <- c("Homo sapiens", "fungi", "mammalia", "Cryptococcus neoformans")
ranks <- c("family", "class", "phylum", "kingdom")
getRanks4Name(nameList, ranks)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
