% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSampling.R
\name{getRepresentative}
\alias{getRepresentative}
\title{Get representative from a list of taxa based on a selected taxonomy rank}
\usage{
getRepresentative(taxonList = NULL, rank = NULL)
}
\arguments{
\item{taxonList}{List of taxa}

\item{rank}{Taxonomy rank used for sub-sampling. Each super taxon of this
rank will have one representative.}
}
\value{
A dataframe contains sub-selected taxa including their IDs, names, as
well as IDs and names of the supertaxa from the selected rank.
}
\description{
Get representative from a list of taxa based on a selected taxonomy rank
}
\examples{
taxonList <- c(
    214684, 246409, 3055, 336722, 3702, 400682, 441375, 45351, 4558, 559292, 
    6239, 7227, "Homo sapiens"
)
getRepresentative(taxonList, rank = "phylum")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
