% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree4Id}
\alias{getTree4Id}
\title{create taxonomy tree for a list of taxon names}
\usage{
getTree4Id(idList = NULL)
}
\arguments{
\item{idList}{list of taxonomy IDs}
}
\value{
A dataframe contains input taxon Ids, their full names and the 
corresponding ranks
}
\description{
create taxonomy tree for a list of taxon names
}
\examples{
idList <- c("9606", "5207", "40674", "4751")
tree <- getTree4Id(idList)
plot(tree)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
