% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree4Name}
\alias{getTree4Name}
\title{create taxonomy tree for a list of taxon names}
\usage{
getTree4Name(nameList = NULL)
}
\arguments{
\item{nameList}{list of taxonomy names}
}
\value{
A dataframe contains input taxon Ids, their full names and the 
corresponding ranks
}
\description{
create taxonomy tree for a list of taxon names
}
\examples{
nameList <- c("Homo sapiens", "fungi", "mammalia", "Cryptococcus neoformans")
tree <- getTree4Name(nameList)
plot(tree)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
