% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idNname.R
\name{id2name}
\alias{id2name}
\title{Get taxon names for a list of taxon IDs}
\usage{
id2name(idList = NULL)
}
\arguments{
\item{idList}{list of taxonomy IDs}
}
\value{
A dataframe contains input taxon Ids and their full names.
}
\description{
Get taxon names for a list of taxon IDs
}
\examples{
idList <- c("9606", "5207", "40674", "4751")
id2name(idList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
