% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{id2rank}
\alias{id2rank}
\title{Get taxonomy ranks for a list of taxon IDs}
\usage{
id2rank(idList = NULL)
}
\arguments{
\item{idList}{list of taxonomy IDs}
}
\value{
A dataframe contains input taxon Ids, their full names and the 
corresponding ranks
}
\description{
Get taxonomy ranks for a list of taxon IDs
}
\examples{
idList <- c("9606", "5207", "40674", "4751")
id2rank(idList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
