% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idNname.R
\name{name2id}
\alias{name2id}
\title{Get taxon IDs for a list of taxon names}
\usage{
name2id(nameList = NULL)
}
\arguments{
\item{nameList}{list of taxonomy names}
}
\value{
A dataframe contains input taxon names and their IDs.
}
\description{
Get taxon IDs for a list of taxon names
}
\examples{
nameList <- c("Homo sapiens", "fungi", "mammalia", "Cryptococcus neoformans")
name2id(nameList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
