% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{name2rank}
\alias{name2rank}
\title{Get taxonomy ranks for a list of taxon names}
\usage{
name2rank(nameList = NULL)
}
\arguments{
\item{nameList}{list of taxonomy names}
}
\value{
A dataframe contains input taxon names, their IDs and the 
corresponding ranks
}
\description{
Get taxonomy ranks for a list of taxon names
}
\examples{
nameList <- c("Homo sapiens", "fungi", "mammalia", "Cryptococcus neoformans")
name2rank(nameList)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
