% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{predictType}
\alias{predictType}
\title{Predict type (ID or NAME) for a list of input taxa}
\usage{
predictType(inputList)
}
\arguments{
\item{inputList}{list of taxa}
}
\value{
ID or NAME
}
\description{
Predict type (ID or NAME) for a list of input taxa
}
\examples{
ids <- c("123", 456, 789)
taxFun:::predictType(ids)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
