% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTaxonomyMatrix.R
\name{sortTaxonomyMatrix}
\alias{sortTaxonomyMatrix}
\title{Sort a taxonomy matrix based on a defined reference species}
\usage{
sortTaxonomyMatrix(taxonList = NULL, refspec = NULL)
}
\arguments{
\item{taxonList}{list of taxonomy IDs}

\item{refspec}{reference taxon}
}
\value{
A dataframe contains clustered taxonomy matrix
}
\description{
Sort a clustered taxonomy matrix for a list of taxon IDs, 
sorted based on a defined reference species
}
\examples{
taxonList <- c("Homo sapiens", "5207", "40674", "4751", "3702", "72664")
sortTaxonomyMatrix(taxonList, "Homo sapiens")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
