% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTaxonomyMatrix.R
\name{sortTaxonomyMatrix4Id}
\alias{sortTaxonomyMatrix4Id}
\title{Create a clustered taxonomy matrix for a list of taxon IDs}
\usage{
sortTaxonomyMatrix4Id(idList = NULL, refspec = NULL)
}
\arguments{
\item{idList}{list of taxonomy IDs}

\item{refspec}{reference taxon}
}
\value{
A dataframe contains clustered taxonomy matrix
}
\description{
Create a clustered taxonomy matrix for a list of taxon IDs, 
sorted based on a defined reference species
}
\examples{
idList <- c("9606", "5207", "40674", "4751", "3702", "72664")
sortTaxonomyMatrix4Id(idList, "Homo sapiens")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
