% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTaxonomyMatrix.R
\name{sortTaxonomyMatrix4Name}
\alias{sortTaxonomyMatrix4Name}
\title{Create a clustered taxonomy matrix for a list of taxon names}
\usage{
sortTaxonomyMatrix4Name(nameList = NULL, refspec = NULL)
}
\arguments{
\item{nameList}{list of taxonomy names}

\item{refspec}{reference taxon}
}
\value{
A dataframe contains clustered taxonomy matrix
}
\description{
Create a clustered taxonomy matrix for a list of taxon names, 
sorted based on a defined reference species
}
\examples{
nameList <- c(
"Homo sapiens", "Mammalia", "Cryptococcus neoformans", "Fungi", 
"Arabidopsis thaliana", "Eutrema salsugineum"
)
sortTaxonomyMatrix4Name(nameList, "Homo sapiens")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
