% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTaxFun.R
\name{taxFun}
\alias{taxFun}
\title{Run taxFun}
\usage{
taxFun(fnName, inputListFile, ranks, refspec, outgroup, outputFile)
}
\arguments{
\item{fnName}{Function name, either "id2name", "name2id", "id2rank",
"name2rank", "getRanks4Id", "getRanks4Name", "sortTaxonomyMatrix4Id",
"sortTaxonomyMatrix4Name", "getTree4Id", "getTree4Name" or
"getAllTaxonomyRanks"}

\item{inputListFile}{Location of input file (list of taxon names or ids)}

\item{ranks}{List of taxonomy ranks of interest (OPTIONAL)}

\item{refspec}{Reference species (REQUIRED for some functions)}

\item{outgroup}{Output species used for rooting the tree (OPTIONAL for tree
functions)}

\item{outputFile}{Location of output file. If not given, it will be saved as
<inputFile.out>}
}
\value{
Output file
}
\description{
Run taxFun
}
\examples{
\donttest{
inputListId <- system.file(
    "extdata", "ids.txt", package = "taxFun", mustWork = TRUE
)
inputListName <- system.file(
    "extdata", "names.txt", package = "taxFun", mustWork = TRUE
)
taxFun("id2name", inputListId)
taxFun("name2rank", inputListName)
taxFun("getRanks4Name", inputListName, ranks = c("class", "phylum"))
}
}
