% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTaxFun.R
\name{taxFun}
\alias{taxFun}
\title{Run taxFun}
\usage{
taxFun(fnName, inputListFile, rank, ranks, refspec, outgroup,
outputFile)
}
\arguments{
\item{fnName}{Function name, either "id2name", "name2id", "getRank", 
"getRanks", "sortTaxonomyMatrix", "createTree", "getRepresentative", 
or "getAllTaxonomyRanks"}

\item{inputListFile}{Location of input file (list of taxon names or ids)}

\item{rank}{A selected taonomy rank (REQUIRED for some functions)}

\item{ranks}{List of taxonomy ranks of interest (OPTIONAL)}

\item{refspec}{Reference species (REQUIRED for some functions)}

\item{outgroup}{Output species used for rooting the tree (OPTIONAL for tree
functions)}

\item{outputFile}{Location of output file. If not given, it will be saved as
<inputFile.out>}
}
\value{
Output file
}
\description{
Run taxFun
}
\examples{
inputListId <- system.file(
    "extdata", "ids.txt", package = "taxFun", mustWork = TRUE
)
inputListName <- system.file(
    "extdata", "names.txt", package = "taxFun", mustWork = TRUE
)
taxFun("id2name", inputListId)
taxFun("getRank", inputListName)
taxFun("getRanks", inputListName, ranks = c("class", "phylum"))
taxFun("createTree", inputListName, outgroup = 9606)
taxFun("sortTaxonomyMatrix", inputListId, refspec = "Homo sapiens")
taxFun("getRepresentative", inputListId, rank = "phylum")
}
