% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTaxFun.R
\name{taxFun}
\alias{taxFun}
\title{Run taxFun}
\usage{
taxFun(fnName, inputTaxaFile, rank, ranks, refspec, supertaxon, 
outgroup, outputFile)
}
\arguments{
\item{fnName}{Function name, either "id2name", "name2id", "getRank", 
"getRanks", "sortTaxonomyMatrix", "createTree", "createTreeSlow",
"getRepresentative", "getClade", "getLCA" or "getAllTaxonomyRanks"}

\item{inputTaxaFile}{Location of input taxa file (taxon names or/and ids)}

\item{rank}{A selected taonomy rank (REQUIRED for some functions)}

\item{ranks}{List of taxonomy ranks of interest (OPTIONAL)}

\item{refspec}{Reference species (REQUIRED for some functions)}

\item{supertaxon}{Supertaxon name or ID (REQUIRED for some functions)}

\item{outgroup}{Output species used for rooting the tree (OPTIONAL for tree
functions)}

\item{outputFile}{Location of output file. If not given, it will be saved as
<inputFile.out>}
}
\value{
Output file
}
\description{
Run taxFun
}
\examples{
inputTaxonIDs <- system.file(
    "extdata", "ids.txt", package = "taxFun", mustWork = TRUE
)
inputTaxonNames <- system.file(
    "extdata", "names.txt", package = "taxFun", mustWork = TRUE
)
taxFun("id2name", inputTaxonIDs)
taxFun("getRank", inputTaxonNames)
taxFun("getRanks", inputTaxonNames, ranks = c("class", "phylum"))
taxFun("createTree", inputTaxonNames, outgroup = 9606)
taxFun("sortTaxonomyMatrix", inputTaxonIDs, refspec = "Homo sapiens")
taxFun("getRepresentative", inputTaxonIDs, rank = "phylum")
taxFun("getClade", inputTaxonIDs, supertaxon = "metazoa")
taxFun("getLCA", inputTaxonNames)
}
